% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba2seasons.R
\docType{data}
\name{nba2seasons}
\alias{nba2seasons}
\title{2 Season's worth of NBA player stats}
\format{
An object of class \code{"data.frame"}
}
\source{
\href{https://www.nba.com/stats/}{NBA Stats}
}
\usage{
data(nba2seasons)
}
\description{
This data was pulled using the (at the time) seemingly error-prone nbaStatR package. I need to revisit but
here it is for testing
}
\examples{
data(nba2seasons)
}
\references{
The NBA
(\href{https://www.nba.com/stats/}{NBA Stats})
}
\keyword{datasets}
