% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allotax_calculations.R
\name{rank_turbulence}
\alias{rank_turbulence}
\title{--------------------------------}
\usage{
rank_turbulence(df1, df2, alpha = 1/3, ties.method = "average")
}
\arguments{
\item{df1}{a frame with name and count columns}

\item{df2}{another frame with name and count columns}

\item{alpha}{alpha as defined in the paper, ranges from zero to Inf}

\item{ties.method}{the method of handling ties, one of the options for the rank function: "average", "first", "last, "random", "max", "min"}
}
\value{
a list containing:
            count_total: counts for the two groups, expressed as percentage of the total count across groups
            items_total: breakdown of items in each group as percentage of items across both groups
            exclusive_items: percent of items in a group exclusive to the group
            normalization: value used to normalize results
            divergences: data frame with columns: name, divergence, sign. Negative sign if item ranked higher in the
                            first group
}
\description{
Calculate values relating to rank turbulence between two groups
}
\examples{

}
